+++ R/checkStuff.R
@@ -111,10 +111,14 @@ checkStuff = function(fun, design, learner, control) {
           ifelse(hasLearnerProperties(learner, "se"), "",
             "\nBut this learner does not support prediction of standard errors!"))
       }
-      if (control$multipoint.method == "cl" && infill.crit.id != "ei")
-        stopf("Multi-point proposal using constant liar needs the infill criterion 'ei' (expected improvement), but you used '%s'!", infill.crit.id)
-      if (control$multipoint.method == "cb" && infill.crit.id != "cb")
+      if (control$multipoint.method == "cl" && infill.crit.id %nin% c("ei", "aei", "cb")) {
+        stopf("Multi-point proposal using constant liar needs the infill criterion 'ei' or 'aei', but you used '%s'!", infill.crit.id)
+      } else if (control$multipoint.method == "cl" && infill.crit.id == c("cb")) {
+        warningf("You are using multi-point proposal with constant liar and infill criterion 'cb'. Are you sure you don't want to use 'cb' as multi-point method?")
+      }
+      if (control$multipoint.method == "cb" && infill.crit.id != "cb") {
         stopf("Multi-point proposal using parallel cb needs the infill criterion 'cb' (confidence bound), but you used '%s'!", infill.crit.id)
+      }
     }
   }
 
@@ -130,8 +134,6 @@ checkStuff = function(fun, design, learner, control) {
     if (control$multiobj.method == "mspot" && control$infill.opt != "nsga2")
       stopf("For multi-objective 'mspot' infil.opt must be set to 'nsga2'!")
   }
-
-  # multifidelity stuff
 
   # propose point filtering
   # FIXME: implement something that works for integer and discrte params